Make separate stats module
