+++ docs/ref/stats.md
@@ -1,0 +1,1 @@
+::: evalio.stats
