+++ docs/ref/types.md
@@ -1,16 +1,1 @@
 ::: evalio.types
-    options:
-        filters:
-            - "!Trajectory"
-            - "!MetricKind"
-            - "!Metric"
-            - "!Error"
-
-::: evalio.types
-    options:
-        show_root_toc_entry: false
-        members:
-            - Trajectory
-            - MetricKind
-            - Metric
-            - Error
