+++ tests/test_stats.py
@@ -1,6 +1,7 @@
 from copy import deepcopy
 from evalio.types import Stamp
 from evalio.types import SE3, Trajectory
+from evalio import stats
 from utils import rand_se3, isclose_se3
 import numpy as np
 
@@ -20,7 +21,7 @@ def test_already_aligned():
 
     traj1_out = deepcopy(traj)
     traj2_out = deepcopy(traj)
-    Trajectory._align_stamps(traj1_out, traj2_out)
+    stats.align_stamps(traj1_out, traj2_out)
 
     assert (traj1_out, traj2_out) == (traj, traj)
 
@@ -39,7 +40,7 @@ def test_subsample_first():
     )
 
     traj1_out, traj2_out = deepcopy(traj1), deepcopy(traj2)
-    Trajectory._align_stamps(traj1_out, traj2_out)
+    stats.align_stamps(traj1_out, traj2_out)
 
     if traj1_out != traj2:
         raise ValueError(
@@ -67,7 +68,7 @@ def test_overstep():
     )
 
     traj1_out, traj2_out = deepcopy(traj1), deepcopy(traj2)
-    Trajectory._align_stamps(traj1_out, traj2_out)
+    stats.align_stamps(traj1_out, traj2_out)
 
     if traj1_out != traj1:
         raise ValueError(
@@ -80,7 +81,7 @@ def test_overstep():
         )
 
 
-def test_align_poses():
+def testalign_poses():
     np.random.seed(0)
     gt = Trajectory(
         metadata={},
@@ -96,7 +97,7 @@ def test_align_poses():
         poses=[offset * pose for pose in gt.poses],
     )
 
-    Trajectory._align_poses(traj2, gt)
+    stats.align_poses(traj2, gt)
 
     for a, b in zip(traj2.poses, gt.poses):
         assert isclose_se3(a, b), f"{a} != {b}"
